+++ src/lib/index.ts
@@ -6,3 +6,7 @@
 export function isString(n: unknown) {
   return typeof n === "string" || n instanceof String;
 }
+
+export function formatImageName(index: number) {
+  return index < 10 ? `0${index}` : `${index}`;
+}